Extract restore backup codec and validation boundaries
